Update to use latest CocoaPods

+++ spec/spec_helper/prepare_spec_repos.rb
@@ -7,6 +7,7 @@ def cocoapods_generate_specs_prepare_spec_repos
   require 'cocoapods/executable'
 
   cocoapods_generate_specs_cp_repos_dir.each_child do |c|
+    next unless File.exist?(c.join('url'))
     FileUtils.rm_rf c.join('.git')
 
     Pod::Executable.capture_command(:git, ['init', c.to_s])
